cleanup: Move php construct to construct.ts
